Broke functions out of PopulateSubmissionViewHolder.java into SubmissionThumbnailHelper.java
