+++ app/src/main/java/me/edgan/redditslide/SubmissionViews/PopulateSubmissionViewHolder.java
@@ -66,7 +66,6 @@ import me.edgan.redditslide.Adapters.SubmissionViewHolder;
 import me.edgan.redditslide.Authentication;
 import me.edgan.redditslide.CommentCacheAsync;
 import me.edgan.redditslide.ContentType;
-import me.edgan.redditslide.DataShare;
 import me.edgan.redditslide.ForceTouch.PeekViewActivity;
 import me.edgan.redditslide.HasSeen;
 import me.edgan.redditslide.Hidden;
@@ -90,13 +89,13 @@ import me.edgan.redditslide.util.BlendModeUtil;
 import me.edgan.redditslide.util.ClipboardUtil;
 import me.edgan.redditslide.util.CompatUtil;
 import me.edgan.redditslide.util.DisplayUtil;
-import me.edgan.redditslide.util.GifUtils;
 import me.edgan.redditslide.util.JsonUtil;
 import me.edgan.redditslide.util.LayoutUtils;
 import me.edgan.redditslide.util.LinkUtil;
 import me.edgan.redditslide.util.NetworkUtil;
 import me.edgan.redditslide.util.OnSingleClickListener;
 import me.edgan.redditslide.util.SubmissionParser;
+import me.edgan.redditslide.util.SubmissionThumbnailHelper;
 
 import net.dean.jraw.ApiException;
 import net.dean.jraw.fluent.FlairReference;
@@ -188,7 +187,7 @@ public class PopulateSubmissionViewHolder {
                                         case DEVIANTART:
                                         case XKCD:
                                         case IMAGE:
-                                            openImage(
+                                            SubmissionThumbnailHelper.openImage(
                                                     type,
                                                     contextActivity,
                                                     submission,
@@ -218,7 +217,7 @@ public class PopulateSubmissionViewHolder {
                                             }
                                             break;
                                         case REDDIT:
-                                            openRedditContent(submission.getUrl(), contextActivity);
+                                            SubmissionThumbnailHelper.openRedditContent(submission.getUrl(), contextActivity);
                                             break;
                                         case REDDIT_GALLERY:
                                             if (SettingValues.album) {
@@ -363,7 +362,7 @@ public class PopulateSubmissionViewHolder {
                                         case VREDDIT_REDIRECT:
                                         case GIF:
                                         case VREDDIT_DIRECT:
-                                            openGif(
+                                            SubmissionThumbnailHelper.openGif(
                                                     contextActivity,
                                                     submission,
                                                     holder.getBindingAdapterPosition());
@@ -401,228 +400,6 @@ public class PopulateSubmissionViewHolder {
                         }
                     }
                 });
-    }
-
-    public static void openRedditContent(String url, Context c) {
-        OpenRedditLink.openUrl(c, url, true);
-    }
-
-    public static void openImage(
-            ContentType.Type type,
-            Activity contextActivity,
-            Submission submission,
-            HeaderImageLinkView baseView,
-            int adapterPosition) {
-        if (SettingValues.image) {
-            Intent myIntent = new Intent(contextActivity, MediaView.class);
-            myIntent.putExtra(MediaView.SUBREDDIT, submission.getSubredditName());
-            myIntent.putExtra(EXTRA_SUBMISSION_TITLE, submission.getTitle());
-            String previewUrl;
-            String url = submission.getUrl();
-
-            if (baseView != null
-                    && baseView.lq
-                    && SettingValues.loadImageLq
-                    && type != ContentType.Type.XKCD) {
-                myIntent.putExtra(MediaView.EXTRA_LQ, true);
-                myIntent.putExtra(MediaView.EXTRA_DISPLAY_URL, baseView.loadedUrl);
-            } else if (submission.getDataNode().has("preview")
-                    && submission
-                            .getDataNode()
-                            .get("preview")
-                            .get("images")
-                            .get(0)
-                            .get("source")
-                            .has("height")
-                    && type
-                            != ContentType.Type
-                                    .XKCD) { // Load the preview image which has probably already
-                // been cached in memory instead of the direct link
-                previewUrl =
-                        submission
-                                .getDataNode()
-                                .get("preview")
-                                .get("images")
-                                .get(0)
-                                .get("source")
-                                .get("url")
-                                .asText();
-                if (baseView == null || (!SettingValues.loadImageLq && baseView.lq)) {
-                    myIntent.putExtra(MediaView.EXTRA_DISPLAY_URL, previewUrl);
-                } else {
-                    myIntent.putExtra(MediaView.EXTRA_DISPLAY_URL, baseView.loadedUrl);
-                }
-            }
-            myIntent.putExtra(MediaView.EXTRA_URL, url);
-            PopulateBase.addAdaptorPosition(myIntent, submission, adapterPosition);
-            myIntent.putExtra(MediaView.EXTRA_SHARE_URL, submission.getUrl());
-
-            contextActivity.startActivity(myIntent);
-
-        } else {
-            LinkUtil.openExternally(submission.getUrl());
-        }
-    }
-
-    public static void openGif(
-            Activity contextActivity, Submission submission, int adapterPosition) {
-        if (SettingValues.gif) {
-            DataShare.sharedSubmission = submission;
-
-            Intent myIntent = new Intent(contextActivity, MediaView.class);
-            myIntent.putExtra(MediaView.SUBREDDIT, submission.getSubredditName());
-            myIntent.putExtra(EXTRA_SUBMISSION_TITLE, submission.getTitle());
-
-            GifUtils.AsyncLoadGif.VideoType t =
-                    GifUtils.AsyncLoadGif.getVideoType(submission.getUrl());
-
-            if (t == GifUtils.AsyncLoadGif.VideoType.VREDDIT) {
-                if (submission.getDataNode().has("media")
-                        && submission.getDataNode().get("media").has("reddit_video")
-                        && submission
-                                .getDataNode()
-                                .get("media")
-                                .get("reddit_video")
-                                .has("hls_url")) {
-                    myIntent.putExtra(
-                            MediaView.EXTRA_URL,
-                            StringEscapeUtils.unescapeJson(
-                                            submission
-                                                    .getDataNode()
-                                                    .get("media")
-                                                    .get("reddit_video")
-                                                    .get("dash_url") // In the future, we could
-                                                    // load the HLS url as well
-                                                    .asText())
-                                    .replace("&amp;", "&"));
-                } else if (submission.getDataNode().has("media")
-                        && submission.getDataNode().get("media").has("reddit_video")) {
-                    myIntent.putExtra(
-                            MediaView.EXTRA_URL,
-                            StringEscapeUtils.unescapeJson(
-                                            submission
-                                                    .getDataNode()
-                                                    .get("media")
-                                                    .get("reddit_video")
-                                                    .get("fallback_url")
-                                                    .asText())
-                                    .replace("&amp;", "&"));
-                } else if (submission.getDataNode().has("crosspost_parent_list")) {
-                    myIntent.putExtra(
-                            MediaView.EXTRA_URL,
-                            StringEscapeUtils.unescapeJson(
-                                            submission
-                                                    .getDataNode()
-                                                    .get("crosspost_parent_list")
-                                                    .get(0)
-                                                    .get("media")
-                                                    .get("reddit_video")
-                                                    .get("dash_url")
-                                                    .asText())
-                                    .replace("&amp;", "&"));
-                } else {
-                    new OpenVRedditTask(contextActivity, submission.getSubredditName())
-                            .executeOnExecutor(AsyncTask.THREAD_POOL_EXECUTOR, submission.getUrl());
-                    return;
-                }
-
-            } else if (t.shouldLoadPreview()
-                    && submission.getDataNode().has("preview")
-                    && submission.getDataNode().get("preview").get("images").get(0).has("variants")
-                    && submission
-                            .getDataNode()
-                            .get("preview")
-                            .get("images")
-                            .get(0)
-                            .get("variants")
-                            .has("mp4")) {
-                myIntent.putExtra(
-                        MediaView.EXTRA_URL,
-                        StringEscapeUtils.unescapeJson(
-                                        submission
-                                                .getDataNode()
-                                                .get("preview")
-                                                .get("images")
-                                                .get(0)
-                                                .get("variants")
-                                                .get("mp4")
-                                                .get("source")
-                                                .get("url")
-                                                .asText())
-                                .replace("&amp;", "&"));
-            } else if (t.shouldLoadPreview()
-                    && submission.getDataNode().has("preview")
-                    && submission
-                            .getDataNode()
-                            .get("preview")
-                            .get("reddit_video_preview")
-                            .has("fallback_url")) {
-                myIntent.putExtra(
-                        MediaView.EXTRA_URL,
-                        StringEscapeUtils.unescapeJson(
-                                        submission
-                                                .getDataNode()
-                                                .get("preview")
-                                                .get("reddit_video_preview")
-                                                .get("fallback_url")
-                                                .asText())
-                                .replace("&amp;", "&"));
-            } else if (t == GifUtils.AsyncLoadGif.VideoType.DIRECT
-                    && submission.getDataNode().has("media")
-                    && submission.getDataNode().get("media").has("reddit_video")
-                    && submission
-                            .getDataNode()
-                            .get("media")
-                            .get("reddit_video")
-                            .has("fallback_url")) {
-                myIntent.putExtra(
-                        MediaView.EXTRA_URL,
-                        StringEscapeUtils.unescapeJson(
-                                        submission
-                                                .getDataNode()
-                                                .get("media")
-                                                .get("reddit_video")
-                                                .get("fallback_url")
-                                                .asText())
-                                .replace("&amp;", "&"));
-
-            } else if (t != GifUtils.AsyncLoadGif.VideoType.OTHER) {
-                myIntent.putExtra(MediaView.EXTRA_URL, submission.getUrl());
-            } else {
-                LinkUtil.openUrl(
-                        submission.getUrl(),
-                        Palette.getColor(submission.getSubredditName()),
-                        contextActivity,
-                        adapterPosition,
-                        submission);
-                return;
-            }
-            if (submission.getDataNode().has("preview")
-                    && submission
-                            .getDataNode()
-                            .get("preview")
-                            .get("images")
-                            .get(0)
-                            .get("source")
-                            .has("height")) { // Load the preview image which has probably
-                // already been cached in memory instead of the
-                // direct link
-                String previewUrl =
-                        submission
-                                .getDataNode()
-                                .get("preview")
-                                .get("images")
-                                .get(0)
-                                .get("source")
-                                .get("url")
-                                .asText();
-                myIntent.putExtra(MediaView.EXTRA_DISPLAY_URL, previewUrl);
-            }
-            PopulateBase.addAdaptorPosition(myIntent, submission, adapterPosition);
-            contextActivity.startActivity(myIntent);
-        } else {
-            LinkUtil.openExternally(submission.getUrl());
-        }
     }
 
     public String reason;
